Date and Time Task 1 is done, Task 2 half
